requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
pypdf>=4.0.0
pyyaml>=6.0
tiktoken>=0.7.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
pytest>=8.0.0
